# docs/openapi/client.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Eau La Maman API",
        "version": "1.0.0",
        "description": "API consomm\u00e9e par l'espace client Nuxt et l'application mobile (livreurs/propri\u00e9taires). Authentification par token Bearer Sanctum \u2014 voir docs/api-auth-contract.md et docs/api-espace-client-contract.md pour les d\u00e9cisions m\u00e9tier et les limites connues que cette sp\u00e9cification ne peut pas exprimer."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local"
        },
        {
            "url": "https://formation.eau-la-maman.com/api",
            "description": "Pr\u00e9prod"
        },
        {
            "url": "https://fello.eau-la-maman.com/api",
            "description": "Production"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/auth/otp-login/request": {
            "post": {
                "operationId": "auth.otp-login.request",
                "tags": [
                    "Autres"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "telephone": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "telephone"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "sent": {
                                            "type": "boolean"
                                        },
                                        "channel": {
                                            "type": "string"
                                        },
                                        "destination_masked": {
                                            "type": "string",
                                            "description": "Ajout\u00e9 le 27/08/2026 (demande front) : la coordonn\u00e9e R\u00c9ELLEMENT\nutilis\u00e9e pour ce canal, d\u00e9j\u00e0 masqu\u00e9e c\u00f4t\u00e9 serveur (cf.\nOtpDestinationMasker) \u2014 jamais l'adresse/le num\u00e9ro complet,\njamais reconstruite c\u00f4t\u00e9 client \u00e0 partir d'une autre source."
                                        },
                                        "cooldown_seconds": {
                                            "type": "integer",
                                            "const": 30
                                        }
                                    },
                                    "required": [
                                        "sent",
                                        "channel",
                                        "destination_masked",
                                        "cooldown_seconds"
                                    ]
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "const": "Impossible d'envoyer un code de connexion \u00e0 ce compte pour le moment."
                                                }
                                            },
                                            "required": [
                                                "error"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "const": "Aucun canal disponible pour recevoir un code de connexion pour le moment."
                                                }
                                            },
                                            "required": [
                                                "error"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "retry_after_seconds": {
                                            "type": "integer"
                                        }
                                    },
                                    "required": [
                                        "error",
                                        "retry_after_seconds"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "const": "Aucun compte trouv\u00e9 pour ce num\u00e9ro de t\u00e9l\u00e9phone."
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                },
                "security": []
            }
        },
...
        "/v1/mobile/web-push/subscriptions": {
            "post": {
                "operationId": "client.web-push.subscriptions.store",
                "summary": "Idempotent : upsert par `endpoint_hash` (unique globalement \u2014 cf.\nmigration). Un m\u00eame endpoint r\u00e9abonn\u00e9 par un autre compte (poste\npartag\u00e9) lui est r\u00e9assign\u00e9, jamais dupliqu\u00e9",
                "tags": [
                    "Autres"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WebPushSubscriptionStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "success"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "client.web-push.subscriptions.destroy",
                "summary": "Ne supprime QUE l'abonnement identifi\u00e9 par `endpoint`, scop\u00e9 au compte\nauthentifi\u00e9 \u2014 jamais un \"delete all\" (un User peut avoir plusieurs\nappareils). Idempotent : un endpoint d\u00e9j\u00e0 absent (ou appartenant \u00e0 un\nautre compte) renvoie le m\u00eame succ\u00e8s, sans jamais confirmer/infirmer\nson existence pour un tiers",
                "tags": [
                    "Autres"
                ],
                "parameters": [
                    {
                        "name": "endpoint",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uri",
                            "maxLength": 2048
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "success"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "description": "Token Sanctum obtenu via `POST /api/auth/login` (champ `token` de la r\u00e9ponse). Envoyer `Authorization: Bearer <token>` sur toute route prot\u00e9g\u00e9e. Coller un token de d\u00e9veloppement ici (bouton Authorize) pour tester les endpoints prot\u00e9g\u00e9s \u2014 jamais un token d'un compte r\u00e9el en production.",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        },
        "schemas": {
...
            "WebPushSubscriptionStoreRequest": {
                "type": "object",
                "description": "Payload standard `PushSubscription.toJSON()` \u2014 cf.\nhttps://developer.mozilla.org/docs/Web/API/PushSubscription/toJSON.\nBornes de taille explicites (point s\u00e9curit\u00e9 du rapport Web Push) :\nun `endpoint`/des cl\u00e9s anormalement longs sont rejet\u00e9s plut\u00f4t\nqu'accept\u00e9s silencieusement.",
                "properties": {
                    "endpoint": {
                        "type": "string",
                        "format": "uri",
                        "maxLength": 2048
                    },
                    "keys": {
                        "type": "object",
                        "properties": {
                            "p256dh": {
                                "type": "string",
                                "maxLength": 255
                            },
                            "auth": {
                                "type": "string",
                                "maxLength": 255
                            }
                        },
                        "required": [
                            "p256dh",
                            "auth"
                        ]
                    },
                    "content_encoding": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "enum": [
                            "aes128gcm",
                            "aesgcm",
                            null
                        ]
                    }
                },
                "required": [
                    "endpoint",
                    "keys"
                ],
                "title": "WebPushSubscriptionStoreRequest"
            }
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}